pyproject.toml
[project]
name = "lyralabs"
version = "0.1.0"
description = "Python-first autonomous AI coworker SaaS (GHL + Google wedge)"
requires-python = ">=3.12,<3.15"
readme = "README.md"
license = { text = "Proprietary" }
authors = [{ name = "Muhammad Sahil" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "email-validator>=2.2.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.13.0",
    "redis>=5.1.0",
    "celery[redis]>=5.4.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "cryptography>=43.0.0",
    "slack-bolt>=1.21.0",
    "slack-sdk>=3.33.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "langgraph>=0.2.40",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "litellm>=1.50.0",
    "anthropic>=0.39.0",
    "stripe>=11.1.0",
    "weasyprint>=62.3",
    "plotly>=5.24.0",
    "kaleido==0.2.1",
    "python-pptx>=1.0.0",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.12",
    "orjson>=3.10.0",
    "qdrant-client>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "respx>=0.21.0",
    "freezegun>=1.5.0",
]
teams = [
    "botbuilder-core>=4.16.0",
    "botbuilder-schema>=4.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["packages/lyra_core", "apps/api", "apps/worker"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "ASYNC", "S", "RUF"]
ignore = ["S101", "E501"]

[tool.ruff.lint.per-file-ignores]
"**/migrations/**" = ["N", "E"]
"**/tests/**" = ["S"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict_optional = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]
